client/requirements.txt
adhawk-ble==0.2
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
bleak==0.21.1
bleak-winrt==1.2.0
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
colorama==0.4.6
frozenlist==1.4.0
google-api-core==2.11.1
google-auth==2.23.0
google-cloud-vision==3.4.3
googleapis-common-protos==1.60.0
grpcio==1.58.0
grpcio-status==1.58.0
idna==3.4
multidict==6.0.4
numpy==1.25.2
openai==0.28.0
opencv-python==4.8.0.76
proto-plus==1.22.3
protobuf==4.24.3
pvcheetah==1.1.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAudio==0.2.13
pydub==0.25.1
python-dotenv==1.0.0
requests==2.31.0
rsa==4.9
SpeechRecognition==3.10.0
tqdm==4.66.1
typing_extensions==4.7.1
urllib3==1.26.16
yarl==1.9.2
